requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-dotenv
flower